+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is a list of changes I've made over the upstream Couscous.
 
+## New
+
+* pageflow and navigation support
+
 ## Fixes
 
 * preserve case of filenames